always cast mp3 bitrates to int, so that CBR and VBR output behaves the same

+++ tinytag/tinytag.py
@@ -568,7 +568,7 @@ class ID3(TinyTag):
                     xframes, byte_count, toc, vbr_scale = ID3._parse_xing_header(fh)
                     if xframes and xframes != 0 and byte_count:
                         self.duration = xframes * ID3.samples_per_frame / float(self.samplerate)
-                        self.bitrate = byte_count * 8 / self.duration / 1000
+                        self.bitrate = int(byte_count * 8 / self.duration / 1000)
                         self.audio_offset = fh.tell()
                         return
                     continue
